ページコンポーネントのTS化
