+++ src/pages/blog/index.tsx
@@ -1,44 +1,55 @@
+import { InferGetStaticPropsType, NextPage } from 'next';
 import { SEO } from '@/components/Seo';
 import { Layout } from '@/components/layout/Layout';
 import { Container } from '@/components/layout/Container';
 import { CategoryHeader } from '@/components/CategoryHeader';
-import { CategoryList } from '@/components/CategoryList';
 import { BlogCard } from '@/components/BlogCard';
 import { MenuButton } from '@/components/layout/MenuButton';
+import { CategoryList } from '@/components/CategoryList';
 import { useToggleMenu } from '@/hooks/useToggleMenu';
+import { getDataForBlogHome } from '@/lib/api';
+import { getCategoryIDs, matchCategories } from '@/lib/utils';
 
-import { getBlogCategorySlug, getDataForCategory } from '@/lib/api';
+type Props = InferGetStaticPropsType<typeof getStaticProps>;
 
-const Category = ({ blogPosts, category, categories, slug, preview }) => {
+const pageMeta = {
+  title: `Blog`,
+  description: `ブログの全記事一覧`,
+  path: '/blog',
+};
+
+const Blog: NextPage<Props> = ({
+  preview,
+  allBlogHomeData: { blogPosts, categories },
+}) => {
   const [isMenuOpen, toggleMenu] = useToggleMenu(false);
-  const { name: categoryName } = category[0];
-  const pageMeta = {
-    title: `${categoryName}カテゴリー記事一覧`,
-    description: `${categoryName}カテゴリーの記事一覧`,
-    path: `/blog/category/${slug}`,
-  };
   return (
     <>
       <SEO meta={pageMeta} />
       <Layout>
         <Container>
-          <CategoryHeader className="mb-5 md:mb-10">
-            {categoryName}
-          </CategoryHeader>
-          <div className="flex gap-10">
+          <CategoryHeader className="mb-5 md:mb-10">ブログ</CategoryHeader>
+          <div className="md:flex gap-10">
             <div className="grow py-5 mx-auto">
               {blogPosts.length ? (
                 <ul className="flex flex-col gap-6">
                   {blogPosts.map(
-                    ({ title, publishDate, slug, sys: { id } }) => (
-                      <BlogCard
-                        key={id}
-                        title={title}
-                        publishDate={publishDate}
-                        slug={slug}
-                        categories={category}
-                      />
-                    )
+                    ({ title, publishDate, slug, sys: { id }, ...post }) => {
+                      const categoryIDs = getCategoryIDs(post);
+                      const matchedCategories = matchCategories(
+                        categoryIDs,
+                        categories
+                      );
+                      return (
+                        <BlogCard
+                          key={id}
+                          title={title}
+                          publishDate={publishDate}
+                          slug={slug}
+                          categories={matchedCategories}
+                        />
+                      );
+                    }
                   )}
                 </ul>
               ) : (
@@ -62,29 +73,11 @@ const Category = ({ blogPosts, category, categories, slug, preview }) => {
   );
 };
 
-export default Category;
+export default Blog;
 
-export async function getStaticProps({ params, preview = false }) {
-  const slug = params.slug;
-  const { blogPosts, category, categories } = await getDataForCategory(
-    slug,
-    preview
-  );
+export const getStaticProps = async ({ preview = false }) => {
+  const allBlogHomeData = await getDataForBlogHome(preview);
   return {
-    props: {
-      preview,
-      blogPosts,
-      category,
-      categories,
-      slug,
-    },
+    props: { preview, allBlogHomeData },
   };
-}
-
-export async function getStaticPaths() {
-  const allCategory = await getBlogCategorySlug();
-  return {
-    paths: allCategory?.map(({ slug }) => `/blog/category/${slug}`) ?? [],
-    fallback: false,
-  };
-}
+};
